:octopus: CLEAN redistribuiting code of commons in map into file per component
